pyproject.toml
[project]
name = "detectcandlestick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "mplfinance>=0.12.10b0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "ta-lib>=0.6.5",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "typer>=0.16.0",
    "yfinance>=0.2.65",
]
